% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmdf.R
\name{epdf}
\alias{epdf}
\title{Estimate empirical probability density function}
\usage{
epdf(x, npos = 1000, scale = 0.1)
}
\arguments{
\item{x}{A vector of data for estimating the pdf.}

\item{npos}{A vector of positions to discrete the pdf.}

\item{scale}{A real value to scale the support of pdf. To keep the accuracy of estimation, it should be less than 0.15.}
}
\value{
A list values (positions, probabilities) of the pdf.
}
\description{
The pdf is estimated from data \code{x}.
}
\examples{
# Generate data
x <- rnorm(1000)

# Estimate the pdf.
res <- epdf(x)
plot(x=res$positions, y=res$probabilities, type="l")

}
