% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epmdf.R
\name{epmf}
\alias{epmf}
\title{Estimate empirical probability mass function}
\usage{
epmf(x)
}
\arguments{
\item{x}{A vector of data for estimating the pmf.}
}
\value{
A dataframe with its first column including all categories and the second column including the empirical pmf.
}
\description{
The pmf is estimated from data \code{x}.
}
\examples{
# Generate data
data(iris)
x <- sample(iris$Species, 20)

# Estimate the pmf.
epmf <- epmf(x)
epmf

}
