% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxResOCI_K.R
\name{maxResOCI_K}
\alias{maxResOCI_K}
\title{Find the best K clustering (partitioning) result from the result of weighted K-Means algorithm}
\usage{
maxResOCI_K(kmeansRes, weighted, a)
}
\arguments{
\item{kmeansRes}{A integer to specify the number of clusters.}

\item{weighted}{A bool value to indicate the K-Means algorithm is weighted or not.}

\item{a}{A real number which is the power parameter of generalized overlapping clustering index. It is used to control the number of clusters. The bigger the \code{a}, the more clusters we get.}
}
\value{
A list to store the best (i.e. the maximum OCI) clustering results among the \code{n_sim} K-Means clustering results.
The list includes the maximum OIC value and the best clustering result which includes the clusters, ams and centroids (check ?KMeansCluster).
}
\description{
Fix the number of clusters K and find the best clustering result from the result of weighted K-Means algorithm.
}
\details{
Fix the number of clusters K and find the best clustering result from the result of weighted K-Means algorithm.
}
\examples{

# Discrete distributions
set.seed(123)
data1 = rbinom(n=50,size=20,prob=0.3)
data2 = rbinom(n=30,size=10,prob=0.5)
data3 = rbinom(n=20,size=15,prob=0.2)
data4 = rbinom(n=28,size=25,prob=0.4)
data5 = rpois(35,4)
data6 = rpois(45,9)
data7 = rpois(15,8)
ds = list(data1,data2,data3,data4,data5,data6,data7)
class(ds)<-"pmdfs"
plot(data=ds,distribution_type="pmf")
resKM = KMeansCluster(2,ds,0)
res = maxResOCI_K(kmeansRes=resKM, weighted=F, a=0.2)
res$maxOIC_K
res$clusterResult$clusters
plot(data=ds,distribution_type="pmf",clusterResult=res$clusterResult)

}
\references{
The weighted K-Means clustering algorithm in the paper
\emph{Distribution-free Overlapping Indices for Dynamic Information Borrowing in Bayesian Hierarchical Modeling}.
}
