% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHMOI_Meta.R
\name{Meta}
\alias{Meta}
\title{BHMOI for Meta-analysis}
\usage{
Meta(
  mus,
  sigmas,
  nSample,
  nGroups,
  a = 0.3,
  weighted = FALSE,
  b = 5,
  alphaY = 1,
  alpha.min = 1,
  betaY = 10,
  alpha.max = 50,
  beta = 10,
  tau0 = 0.01,
  mu0 = .logit(0.1),
  K = 0,
  fast = 1,
  n_sim = 3,
  seed = NULL,
  nIter = 2000,
  kmIters = 50,
  grid = 1000
)
}
\arguments{
\item{mus}{A vector includes the sample means of studies.}

\item{sigmas}{A vector includes the sample standard devistions of studies.}

\item{nSample}{A integer denotes the sample size that you want to draw from every study.}

\item{nGroups}{A integer denotes the number of studies.}

\item{a}{A real number in the interval \link{0,1}. It is used to control the number of clusters. The bigger the \code{a}, the more clusters we get.}

\item{weighted}{A bool value to indicate the K-Means algorithm is weighted (include parameter b) or not.}

\item{b}{A real number which is the power parameter of generalized weighted K-Means algorithm.}

\item{alphaY}{A real value denotes the parameter of the gamma prior for the precision of the real data.}

\item{alpha.min}{A real value denotes the parameter that correspond approximately to no borrowing in the gamma prior.}

\item{betaY}{A real value denotes the parameter of the gamma prior for the precision of the real data.}

\item{alpha.max}{A real value denotes the parameter that correspond approximately to full borrowing in the gamma prior.}

\item{beta}{A real value denotes the parameter in the gamma prior.}

\item{tau0}{A real value denotes the precision of the hyper-parameter of the cluster mean.}

\item{mu0}{A real value denotes the mean of the hyper-parameter of the cluster mean.}

\item{K}{A integer to set the number of cluster.}

\item{fast}{A bool value to select the strategy to calculate the next level of K. 1: stop at the first time K does not increase; 0: stop at OCI_K>OCI_{K+1}.}

\item{n_sim}{A integer to set the repeat times of weighted K-Means algorithm. K-Means algorithm can stuck at a local optimal. Increase n_sim can increase the probability to find the global optimal.}

\item{seed}{A integer to set the number of threads for computation.}

\item{nIter}{A integer to set the number of MCMC iterations.}

\item{kmIters}{A integer to set the max number of iterations in K-Means algorithm.}

\item{grid}{A integer the number of discrete points to calculate the probability density.}
}
\value{
A list that stores the results of Bayesian hierarchical model.
}
\description{
BHMOI for Meta-analysis
}
\details{
There are a set datasets in \code{dataSets}. Each dataset includes discrete data points of a distribution. This function will search the best clustering result of the distributions. The number of clusters can be influenced by parameter \code{a}. The bigger the \code{a}, the more clusters we get.
}
\examples{

mus = c(0.7091362, 0.3548641, 1.7911700, 0.1824552, 0.4218509, 0.6300000, 0.7248838, 0.5286638, 0.2840000,
         1.2750682, 0.1036082, 0.3883906, 0.5407398, 0.4261593, 0.5153969, 1.4797260, 0.6125782, 0.6000000)
nGroups = length(mus)
sigmas = c(0.2608202, 0.1963624, 0.3455692, 0.1177874, 0.1448128, 0.1960000, 0.2246641, 0.2104609, 0.1680000,
           0.3371997, 0.1947275, 0.2307689, 0.2443133, 0.2579379, 0.3512737, 0.3152817, 0.2266834, 0.2490000)
res = Meta(mus, sigmas, nGroups)
plot(data=res$initPost,distribution_type="pdf",ylim = c(0,9))
plot(data=res$initPost,distribution_type="pdf",clusterResult=res$maxOCIres$clusterResult,ylim = c(0,9))
plot(data=res$borrowedPost,distribution_type="pdf",clusterResult=res$maxOCIres$clusterResult,ylim = c(0,9))

}
\references{
The paper
\emph{Distribution-free Overlapping Indices for Efficient and Robust Information Borrowing in Bayesian Hierarchical Modeling}.
}
