% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBI.R
\name{OBI}
\alias{OBI}
\title{Calculate OBIs with the result of clustering}
\usage{
OBI(clusterResult)
}
\arguments{
\item{clusterResult}{A list to store the clustering results (see the return in ?KMeansCluster).}
}
\value{
A vector values of OBIs.
}
\description{
This function calculate OBIs with the result of clustering.
}
\details{
This function calculate OBIs with the result of clustering.
}
\examples{

# Discrete distributions
set.seed(123)
data1 = rbinom(n=50,size=20,prob=0.3)
data2 = rbinom(n=30,size=10,prob=0.5)
data3 = rbinom(n=20,size=15,prob=0.2)
data4 = rbinom(n=28,size=25,prob=0.4)
data5 = rpois(35,4)
data6 = rpois(45,9)
data7 = rpois(15,8)
ds = list(data1,data2,data3,data4,data5,data6,data7)
res = maxOCI_K(K=2, dataSets=ds, data_type=0, a=0.2)
res$clusterResult$clusters
OBI(res$clusterResult)

}
